Cleaned up some annoying warnings.
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15170 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_fluid/fluid_volumetric_surfaceFlow_condition.cpp
@@ -1065,7 +1065,7 @@ void FLD::UTILS::FluidVolumetricSurfaceFlowBc::EvaluateVelocities(
   //--------------------------------------------------------------------
   // get the processor rank
   //--------------------------------------------------------------------
-  //int myrank = discret_->Comm().MyPID();
+  // int myrank = discret_->Comm().MyPID();
 
   double time_in_a_period = fmod(time,period_);//time - period_*floor(time/period_);
   // get the flowrate position
@@ -1283,7 +1283,7 @@ void FLD::UTILS::FluidVolumetricSurfaceFlowBc::Velocities(
   int cycle_num = 0;
   double in_cycle_time = 0.0;
   double time2= time- dta_;
-  cycle_num = double(floor(time2/period));
+  cycle_num = int(floor(time2/period));
   in_cycle_time = time2 - period*double(cycle_num);
   flowratespos_ = int((in_cycle_time/dta_));
